# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Personal Operations Manager",
    "description": "A private workspace with review-before-save agent proposals.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "Tasks",
        "operationId": "tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "completed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "low",
                    "medium",
                    "high"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "overdue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Overdue"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TaskResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      },
      "post": {
        "tags": [
          "Workspace"
        ],
        "summary": "Add Task",
        "operationId": "add_task_api_v1_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the csrf_token returned by /auth/session. Never store the session cookie in JavaScript."
          }
        ]
      }
    },
    "/api/v1/tasks/{record_id}": {
      "patch": {
        "tags": [
          "Workspace"
        ],
        "summary": "Edit Task",
        "operationId": "edit_task_api_v1_tasks__record_id__patch",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the csrf_token returned by /auth/session. Never store the session cookie in JavaScript."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Workspace"
        ],
        "summary": "Delete Task",
        "operationId": "delete_task_api_v1_tasks__record_id__delete",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Version"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the csrf_token returned by /auth/session. Never store the session cookie in JavaScript."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      }
    },
    "/api/v1/notes": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "Notes",
        "operationId": "notes_api_v1_notes_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NoteResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      },
      "post": {
        "tags": [
          "Workspace"
        ],
        "summary": "Add Note",
        "operationId": "add_note_api_v1_notes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Use the csrf_token returned by /auth/session. Never store the session cookie in JavaScript."
          }
        ]
      }
    },
...
    "/api/v1/dashboard": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_v1_dashboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sign-in required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid CSRF token or origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict or stale version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "422": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (see Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafeError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Page_NoteResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NoteResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[NoteResponse]"
      },
...
      "Page_TaskResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[TaskResponse]"
      },
...
      "SafeError": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "loc": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "integer"
                      }
                    ]
                  }
                },
                "msg": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "CookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aiops_session"
      }
    }
  }
}